test(cache): split cache contract coverage
